Change type of timestamp col of location history table from INT to REAL

+++ schema.sql
@@ -22,7 +22,7 @@ CREATE TABLE map_cache(
 CREATE TABLE location_history(
     user_id     INT NOT NULL,
     place_id    TEXT NOT NULL,
-    timestamp   INT NOT NULL,
+    timestamp   REAL NOT NULL,
     FOREIGN KEY(user_id) REFERENCES users(user_id),
     FOREIGN KEY(place_id) REFERENCES map_cache(place_id),
     PRIMARY KEY (user_id, place_id)
